photo_circuit_api/requirements.txt
langchain==0.2.7
langchain-community==0.2.7
langchain-openai==0.1.16
openai==1.35.13
python-dotenv==1.0.1
pillow==10.4.0
Jinja2==3.1.4
